Fix test data generation
Position test data records currently have wrong identifiers because the
data is generated in a callback that will not have correct value of the
iterator variable i when it is executed. The fix use a self-called
anonymous function that encapsulates the actual callback function and
therefor saves the state of i in a separate copy.